requirements.txt
<<<<<<< HEAD
absl-py==0.5.0
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==2.0.6
aText==0.8
attrs==18.1.0
autograd==1.3
awscli==1.16.266
Babel==2.5.3
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
backports.weakref==1.0.post1
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
botocore==1.13.2
Bottleneck==1.2.1
cachetools==3.1.1
cdecimal==2.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
clean-text==0.1.1
click==6.7
click-plugins==1.0.4
cligj==0.5.0
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
conda==4.5.11
conda-build==3.10.5
conda-verify==2.0.0
config==0.4.0
configparser==3.5.0
contextlib2==0.5.5
coverage==4.5.4
cryptography==2.2.2
cycler==0.10.0
cymem==2.0.2
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
descartes==1.1.0
dill==0.2.8.2
distributed==1.21.8
docutils==0.14
dominate==2.3.4
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Fiona==1.8.0
Flask==1.0.2
Flask-Assets==0.12
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.3
Flask-GoogleMaps==0.2.6
Flask-Material==0.1.1
Flask-PyMongo==2.3.0
Flask-SocketIO==3.0.2
ftfy==4.4.3
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gast==0.2.0
geopandas==0.4.0
gevent==1.3.0
glob2==0.6
gmpy2==2.0.8
google==2.0.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
greenlet==0.4.13
grin==1.2.1
grpcio==1.15.0
guidedlda==2.0.0.dev22
gunicorn==19.9.0
h2o==3.20.0.8
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.6
ijson==2.4
imageio==2.3.0
imagesize==1.0.0
imbalanced-learn==0.3.3
imblearn==0.0
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.7.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyter-http-over-ws==0.0.6
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kaggle==1.5.6
Keras==2.2.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Lifetimes==0.11.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Markdown==3.0
MarkupSafe==1.0
matplotlib==2.2.2
matplotlib-venn==0.11.5
mccabe==0.6.1
mistune==0.8.3
mizani==0.4.6
mock==2.0.0
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
msgpack-python==0.5.6
multipledispatch==0.5.0
munch==2.3.2
murmurhash==1.0.1
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.0.4
nose==1.3.7
notebook==5.5.0
numba==0.38.0+0.g2a2b772.dirty
numexpr==2.6.5
numpy==1.15.4
numpydoc==0.8.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
palettable==3.1.1
pandas==0.24.2
pandas-datareader==0.7.0
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib==1.0.1
pathlib2==2.3.2
patsy==0.5.0
pbr==4.2.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
plac==0.9.6
plotly==4.1.1
plotly-express==0.4.1
plotnine==0.4.0
pluggy==0.6.0
ply==3.11
pprintpp==0.4.0
preshed==2.0.1
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.5
ptyprocess==0.5.2
py==1.5.3
py4j==0.10.7
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAudio==0.2.7
pycodestyle==2.4.0
pycosat==0.6.3
pycountry==19.8.18
pycountry-convert==0.7.2
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
pygeocoder==1.2.5
Pygments==2.2.0
pylint==1.8.4
pymongo==3.8.0
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyproj==1.9.5.1
PySocks==1.6.8
pyspark==2.4.4
pysqlite==2.8.3
pytest==3.5.1
pytest-cov==2.7.1
pytest-mock==1.11.0
python-crfsuite==0.9.6
python-dateutil==2.7.3
python-docx==0.8.10
python-engineio==3.0.0
python-slugify==4.0.0
python-socketio==2.1.0
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzipcode==1.0
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
regex==2018.1.10
repoze.lru==0.7
requests==2.18.4
retrying==1.3.3
rope==0.10.7
rsa==3.4.2
ruamel-yaml==0.15.35
s3transfer==0.2.1
scandir==1.7
scikit-image==0.13.1
scikit-learn==0.19.1
scikit-multilearn==0.1.0
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
spacy==2.0.16
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
subprocess32==3.5.0
sympy==1.1.1
tables==3.4.3
tabulate==0.8.2
tblib==1.3.2
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.3
textblob==0.15.2
thinc==6.12.0
toolz==0.9.0
torch==1.3.1
tornado==5.0.2
tqdm==4.28.1
traitlets==4.3.2
typing==3.6.4
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.1.0
visitor==0.1.3
wcwidth==0.1.7
webassets==0.12.1
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wordcloud==1.5.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.2.0
zh-core-web-sm==2.0.5
zict==0.1.3
=======
appnope==0.1.0
astroid==2.3.3
astropy==4.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
bokeh==1.4.0
branca==0.4.0
certifi==2019.9.11
chardet==3.0.4
ChatterBot==1.0.5
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
confuse==1.0.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Fiona==1.8.13
folium==0.10.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopandas==0.6.3
geopy==1.21.0
htmlmin==0.1.12
idna==2.8
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==1.5.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kaggle==1.5.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.1.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.2.0
munch==2.5.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numba==0.48.0
numpy==1.17.4
packaging==20.1
pandas==0.25.3
pandas-profiling==2.4.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
phik==0.9.9
pickleshare==0.7.5
Pillow==6.2.1
Pint==0.9
plac==0.9.6
plotly==4.5.0
pluggy==0.13.1
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.1
pygeocoder==1.2.5
Pygments==2.4.2
pylint==2.4.4
pymongo==3.9.0
pyparsing==2.4.6
pyproj==2.4.2.post1
pyrsistent==0.15.7
pytest==5.3.5
pytest-pylint==0.14.1
python-dateutil==2.7.5
python-googlegeocoder==0.4.0
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
ratelim==0.1.6
requests==2.22.0
retrying==1.3.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.12.0
sklearn==0.0
spacy==2.1.9
SQLAlchemy==1.2.19
srsly==0.2.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc==7.0.8
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.24.3
wasabi==0.4.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.2.0
>>>>>>> dc0794fba338efe7a596a25950f8295ee215c98e
